Defend Workflows Cypress CI optimization (#254117)
## Summary

Reduces Security Solution Cypress CI time through spec splits,
load-balanced chunking, deduplication, and increased parallelism across
all Security Solution Cypress jobs.

## Changes

### 1. Load-balanced chunking (`scripts/run_cypress/`)

- **`utils.ts`**: Added `getSpecFileWeight()` (counts `it(` /
`it.skip(`) and `orderSpecFilesForLoadBalance()` (sorts by weight desc,
path tiebreaker).
- **`parallel.ts`** and **`parallel_serverless.ts`**: Order resolved
spec paths before `retrieveIntegrations()` so round-robin gives each
agent a similar mix of heavy/light specs.

### 2. Policy details split (`e2e/policy/`)

- **`policy_details_helpers.ts`**: Extracted shared helpers
(`getAutomaticUpdatesToggle`, `loadProtectionUpdatesUrl`, etc.).
- **`policy_details_protection_updates.cy.ts`**: All "Protection
updates" tests.
- **`policy_details_policy_settings.cy.ts`**: All "Policy settings"
tests.
- Removed single `policy_details.cy.ts`.

### 3. Complete with Endpoint roles split (`e2e/serverless/roles/`)

- Replaced `complete_with_endpoint_roles.cy.ts` (57 tests) with **7
role-specific specs**: `t1_t2_analyst`, `t3_analyst`,
`threat_intelligence_analyst`, `rule_author`, `soc_manager`,
`endpoint_operations_analyst`,
`platform_engineer_endpoint_policy_manager`.

### 4. Essentials with Endpoint roles split (`e2e/serverless/roles/`)

- Replaced `essentials_with_endpoint.roles.cy.ts` (36 tests) with **6
role-specific specs** (same pattern). All remain `describe.skip` per
existing flaky issue.

### 5. RBAC endpoint_list spec splits (`e2e/endpoint_list/`)

- Split combined SIEM version files into **per-version specs** for
better CI distribution:
- `endpoints_rbac_mocked_data_hosts_exist.cy.ts` → 5 files (`_siem_v1`
through `_siem_v5`)
  - `endpoints_rbac_mocked_data_empty_state.cy.ts` → 5 files
  - `endpoints_rbac_mocked_data_policies_exist.cy.ts` → 5 files

### 6. Navigation RBAC spec splits (`e2e/rbac/`)

- Split `navigation.cy.ts` into **6 files**: `navigation_ess_siem_v1`
through `_v5` + `navigation_serverless`.

### 7. Shared RBAC test helpers (deduplication)

Extracted duplicated test logic from 20 per-version spec files into **4
shared helper modules**, reducing ~1,500 lines of code duplication:
- `rbac_hosts_exist_test_suite.ts` — endpoint host setup hooks +
privilege assertion loops
- `rbac_empty_state_test_suite.ts` — privilege loops + onboarding screen
assertions
- `rbac_policies_exist_test_suite.ts` — policy setup hooks + privilege
assertion loops
- `navigation_rbac_test_suite.ts` — shared `allPages` definition + ESS
navigation test suite

### 8. Pipeline parallelism increases

**Defend Workflows** (PR pipeline):
| Job | Before | After |
|-----|--------|-------|
| Defend Workflows Cypress (stateful) | 22 | 24 |
| Defend Workflows Cypress (serverless) | 14 | 16 |

**Other Security Solution Cypress jobs** (bumped +1 across all pipeline
files: PR, `on_merge_fanout`, `chrome_forward_testing`,
`node_pointer_compression`, `node_glibc_217`,
`verify_es_serverless_image`):

| Job | Before | After |
|-----|--------|-------|
| Investigations (stateful) | 8 | 9 |
| Serverless Investigations | 10 | 11 |
| Detection Engine (stateful) | 7 | 8 |
| Rule Management (stateful) | 4 | 5 |
| Serverless Rule Management | 5 | 6 |
| Serverless Rule Mgt — Prebuilt Rules Installation | 1 | 2 |
| Serverless Rule Mgt — Prebuilt Rules Upgrade | 1 | 2 |
| Serverless Explore | 2 | 3 |
| Serverless AI Assistant | 1 | 2 |

## Testing

- No product behavior change; only test layout, shared helpers, and CI
configuration.
- Defend Workflows Cypress (stateful + serverless) should be run to
confirm all scenarios pass.
- Other Security Solution Cypress jobs affected by parallelism changes
should also pass.

---------

(cherry picked from commit 987fc9142958003008f71dace55082f516353220)

# Conflicts:
#	.buildkite/pipelines/chrome_forward_testing.yml
#	.buildkite/pipelines/node_glibc_217.yml
#	.buildkite/pipelines/on_merge_fanout.yml
#	.buildkite/pipelines/pointer_compression.yml
#	.buildkite/pipelines/pull_request/security_solution/ai_assistant.yml
#	.buildkite/pipelines/pull_request/security_solution/defend_workflows.yml
#	.buildkite/pipelines/pull_request/security_solution/detection_engine.yml
#	.buildkite/pipelines/pull_request/security_solution/explore.yml
#	.buildkite/pipelines/pull_request/security_solution/investigations.yml
#	.buildkite/pipelines/pull_request/security_solution/rule_management.yml
