+++ x-pack/solutions/security/plugins/security_solution/public/management/cypress/e2e/endpoint_list/endpoints_rbac_mocked_data_hosts_exist_siem_v2.cy.ts
@@ -1,0 +1,12 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { createRbacHostsExistSuite } from './rbac_hosts_exist_test_suite';
+
+describe('Endpoints page RBAC - some hosts are enrolled (siem v2)', { tags: ['@ess'] }, () => {
+  createRbacHostsExistSuite('siemV2');
+});
