+++ x-pack/solutions/security/plugins/security_solution/public/management/cypress/e2e/endpoint_list/endpoints_rbac_mocked_data_policies_exist_siem_v4.cy.ts
@@ -1,0 +1,16 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { createRbacPoliciesExistSuite } from './rbac_policies_exist_test_suite';
+
+describe(
+  'Endpoints page RBAC - Defend policy is present, but no hosts (siem v4)',
+  { tags: ['@ess'] },
+  () => {
+    createRbacPoliciesExistSuite('siemV4');
+  }
+);
