+++ x-pack/solutions/security/plugins/security_solution/public/management/cypress/e2e/rbac/navigation_ess_siem_v5.cy.ts
@@ -1,0 +1,12 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { createNavigationEssSuite } from './navigation_rbac_test_suite';
+
+describe('Navigation RBAC - ESS (siem v5)', { tags: ['@ess'] }, () => {
+  createNavigationEssSuite('siemV5');
+});
