+++ x-pack/solutions/security/plugins/security_solution/public/management/cypress/e2e/serverless/roles/complete_with_endpoint_roles_rule_author.cy.ts
@@ -1,0 +1,104 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import type { CyIndexEndpointHosts } from '../../../tasks/index_endpoint_hosts';
+import { indexEndpointHosts } from '../../../tasks/index_endpoint_hosts';
+import { login, ROLE } from '../../../tasks/login';
+import type { EndpointArtifactPageId } from '../../../screens';
+import {
+  ensureArtifactPageAuthzAccess,
+  ensureEndpointListPageAuthzAccess,
+  ensureFleetPermissionDeniedScreen,
+  getEndpointManagementPageMap,
+  getNoPrivilegesPage,
+  openRowActionMenu,
+  visitEndpointList,
+  visitFleetAgentList,
+  visitPolicyList,
+} from '../../../screens';
+
+const pageById = getEndpointManagementPageMap();
+
+describe(
+  'User Roles for Security Complete PLI with Endpoint Complete addon - rule author',
+  {
+    tags: ['@serverless', '@skipInServerlessMKI'],
+    env: {
+      ftrConfig: {
+        productTypes: [
+          { product_line: 'security', product_tier: 'complete' },
+          { product_line: 'endpoint', product_tier: 'complete' },
+        ],
+      },
+    },
+  },
+  () => {
+    let loadedEndpoints: CyIndexEndpointHosts;
+
+    before(() => {
+      indexEndpointHosts().then((response) => {
+        loadedEndpoints = response;
+      });
+    });
+
+    after(() => {
+      if (loadedEndpoints) {
+        loadedEndpoints.cleanup();
+      }
+    });
+
+    describe('for role: rule_author', () => {
+      const artifactPagesFullAccess = [
+        pageById.trustedApps,
+        pageById.trustedDevices,
+        pageById.eventFilters,
+        pageById.blocklist,
+      ];
+
+      beforeEach(() => {
+        login(ROLE.rule_author);
+      });
+
+      for (const { id, title } of artifactPagesFullAccess) {
+        it(`should have CRUD access to: ${title}`, () => {
+          ensureArtifactPageAuthzAccess('all', id as EndpointArtifactPageId);
+        });
+      }
+
+      it('should have access to Endpoint list page', () => {
+        ensureEndpointListPageAuthzAccess('all', true);
+      });
+
+      it('should have access to policy management', () => {
+        visitPolicyList();
+        getNoPrivilegesPage().should('not.exist');
+      });
+
+      it(`should have Read access only to: Host Isolation Exceptions`, () => {
+        ensureArtifactPageAuthzAccess(
+          'read',
+          pageById.hostIsolationExceptions.id as EndpointArtifactPageId
+        );
+      });
+
+      it('should NOT have access to Fleet', () => {
+        visitFleetAgentList();
+        ensureFleetPermissionDeniedScreen();
+      });
+
+      it('should have access to Response Actions Log', () => {
+        cy.visit(pageById.responseActionLog);
+        getNoPrivilegesPage().should('not.exist');
+      });
+
+      it('should NOT have access to execute response actions', () => {
+        visitEndpointList();
+        openRowActionMenu().findByTestSubj('console').should('not.exist');
+      });
+    });
+  }
+);
