+++ x-pack/solutions/security/plugins/security_solution/public/management/cypress/e2e/serverless/roles/essentials_with_endpoint_roles_threat_intelligence_analyst.cy.ts
@@ -1,0 +1,69 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { login, ROLE } from '../../../tasks/login';
+import type { EndpointArtifactPageId } from '../../../screens';
+import {
+  ensureEndpointListPageAuthzAccess,
+  getArtifactListEmptyStateAddButton,
+  getEndpointManagementPageList,
+  getEndpointManagementPageMap,
+  getNoPrivilegesPage,
+  ensureFleetPermissionDeniedScreen,
+  visitFleetAgentList,
+} from '../../../screens';
+
+const allPages = getEndpointManagementPageList();
+const pageById = getEndpointManagementPageMap();
+
+// FLAKY: https://github.com/elastic/kibana/issues/170985
+describe.skip(
+  'Roles for Security Essential PLI with Endpoint Essentials addon - threat intelligence analyst',
+  {
+    tags: ['@serverless', '@skipInServerlessMKI'],
+    env: {
+      ftrConfig: {
+        productTypes: [
+          { product_line: 'security', product_tier: 'essentials' },
+          { product_line: 'endpoint', product_tier: 'essentials' },
+        ],
+      },
+    },
+  },
+  () => {
+    describe('for role: threat_intelligence_analyst', () => {
+      const deniedPages = allPages.filter(({ id }) => id !== 'blocklist' && id !== 'endpointList');
+
+      beforeEach(() => {
+        login(ROLE.threat_intelligence_analyst);
+      });
+
+      it('should have access to Endpoint list page', () => {
+        ensureEndpointListPageAuthzAccess('read', true);
+      });
+
+      it(`should have ALL access to: Blocklist`, () => {
+        cy.visit(pageById.blocklist.url);
+        getArtifactListEmptyStateAddButton(pageById.blocklist.id as EndpointArtifactPageId).should(
+          'exist'
+        );
+      });
+
+      for (const { url, title } of deniedPages) {
+        it(`should NOT have access to: ${title}`, () => {
+          cy.visit(url);
+          getNoPrivilegesPage().should('exist');
+        });
+      }
+
+      it('should NOT have access to Fleet', () => {
+        visitFleetAgentList();
+        ensureFleetPermissionDeniedScreen();
+      });
+    });
+  }
+);
