+++ x-pack/solutions/security/plugins/security_solution/scripts/run_cypress/parallel.ts
@@ -32,7 +32,12 @@ import { createKbnClient } from '../endpoint/common/stack_services';
 import type { StartedFleetServer } from '../endpoint/common/fleet_server/fleet_server_services';
 import { startFleetServer } from '../endpoint/common/fleet_server/fleet_server_services';
 import { renderSummaryTable } from './print_run';
-import { parseTestFileConfig, retrieveIntegrations, setDefaultToolingLoggingLevel } from './utils';
+import {
+  orderSpecFilesForLoadBalance,
+  parseTestFileConfig,
+  retrieveIntegrations,
+  setDefaultToolingLoggingLevel,
+} from './utils';
 import { getFTRConfig } from './get_ftr_config';
 
 export const cli = () => {
@@ -134,7 +139,8 @@ ${JSON.stringify(cypressConfigFile, null, 2)}
               : undefined
           ); // convert the glob pattern to concrete file paths
 
-      let files = retrieveIntegrations(concreteFilePaths);
+      const orderedFilePaths = orderSpecFilesForLoadBalance(concreteFilePaths);
+      let files = retrieveIntegrations(orderedFilePaths);
 
       log.info('Resolved spec files after retrieveIntegrations:', files);
